pyproject.toml
[project]
name = "pylora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyinstaller>=6.19.0",
    "pyserial>=3.5",
    "requests>=2.32.5",
    "toml>=0.10.2",
]
